fix: run Jest as command rather than using the outdated gulp-just helper
Officially, Jest doesn't support running from NodeJS (i.e. there is no 
official documentation) and the gulp-jest helper does not work with the 
latest release of jest-cli. Therefore, I decided to run jest as a 
command anyway. This is easy because I was already using a configuration 
file for Jest anyway.