feat: add Dockerfile
Add a Dockerfile to be able to run serve the portfolio from a docker
container. The Dockerfile defines a docker image for NodeJS with gulp
installed globally (i.e. `npm install gulp --global`) that will execute
the command `gulp serve` when run.

In addition, a .dockeringore file was added which exclude NodeJS files
and potentially generated files from the docker image.