Use `csp_evaluator` to continuously lint the CSP policy (#97)
Add the `csp_evaluator` package [1] as a dependency and use it in a
project-specific CLI to (continuously) "lint" (or verify if you will)
the CSP used in this project. To support this, the CSP has been
extracted from the Pug source code into a text file. This allows the
"linter" to easily access (and thus evaluate) it and enables the Pug
templating engine to inject the CSP into the HTML at build time.

One exception is build into the CLI which is for the informative warning
about the `'self'` used in the script-src directive (which allows for
scripts originating from the website itself) because we do not consider
this a problem for this project (i.e. none of "JSONP, AngularJS or user
uploaded files" apply).

This commit also simplified the CSP in a manner that is compatible with
the existing website and passes the `csp_evaluator`.

--
1. https://www.npmjs.com/package/csp_evaluator