Harden GitHub Actions workflows (#111)
Update all GitHub Actions workflows following an analysis by Zizmor
(https://github.com/woodruffw/zizmor). In particular, this avoids
persisting git credentials when the job does not need it (which I
believe is all jobs).

Zizmor did have one more concern - overly permissive `permissions:
read-all` - but this was not addressed because I think this is okay,
the project is entirely open so I don't see a risk of an attacker
reading anything.