Replace `anywhere` with `http-server` (#113)
Remove the dependency on the outdated and unmaintained `anywhere` and
replace it with the more up-to-date `http-server`.

This adds an exception for the `union` package to the licensee config.
This package is introduced transitively by `http-server` and lacks
license information in its manifest. Its LICENSE file and README specify
the license as being MIT.